% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stats.R, R/old_codes.R
\name{add_fit_cri_bayes}
\alias{add_fit_cri_bayes}
\title{Bayesian CrIs for infit & outfit using draws stored in residuals}
\usage{
add_fit_cri_bayes(fit_stats, res_df, facet, prob = 0.95)

add_fit_cri_bayes(fit_stats, res_df, facet, prob = 0.95)
}
\arguments{
\item{fit_stats}{tibble from facet_infit_outfit()}

\item{res_df}{tibble from extract_residuals() with draws}

\item{facet}{facet name}

\item{prob}{probability for credible intervals}
}
\value{
fit_stats tibble with added CrI columns
}
\description{
Bayesian CrIs for infit & outfit using draws stored in residuals

Bayesian CrIs for infit & outfit using draws stored in residuals
}
