% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stats.R, R/old_codes.R
\name{add_zstd}
\alias{add_zstd}
\title{Add approximate z-standardized infit/outfit
(Wright & Masters cube-root transformation, approximate df = n_obs - 1)}
\usage{
add_zstd(fit_stats)

add_zstd(fit_stats)
}
\arguments{
\item{fit_stats}{tibble from facet_infit_outfit()}
}
\value{
fit_stats tibble with added z-standardized columns
}
\description{
Add approximate z-standardized infit/outfit
(Wright & Masters cube-root transformation, approximate df = n_obs - 1)

Add approximate z-standardized infit/outfit
(Wright & Masters cube-root transformation, approximate df = n_obs - 1)
}
