% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_generation.R
\name{build_stan_code}
\alias{build_stan_code}
\title{Build Stan code from specification and priors}
\usage{
build_stan_code(spec, priors)
}
\arguments{
\item{spec}{bmfrm_spec object from parse_bmfrm_formula()}

\item{priors}{list of bmfrm_prior objects}
}
\value{
character string with complete Stan code
}
\description{
Build Stan code from specification and priors
}
