% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmfrm.R
\name{check_convergence}
\alias{check_convergence}
\title{Check MCMC convergence}
\usage{
check_convergence(fit)
}
\arguments{
\item{fit}{cmdstanr fit object}
}
\value{
list with convergence information
}
\description{
Check MCMC convergence
}
\keyword{internal}
