% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{check_data_suitability}
\alias{check_data_suitability}
\title{Check data suitability for MFRM}
\usage{
check_data_suitability(data, response, facets)
}
\arguments{
\item{data}{data.frame}

\item{response}{response column name}

\item{facets}{facet column names}
}
\value{
list of diagnostic information
}
\description{
Check data suitability for MFRM
}
