% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{clean_model_cache}
\alias{clean_model_cache}
\title{Clean old compiled models from cache}
\usage{
clean_model_cache(cache_dir, max_age_days = 30, dry_run = FALSE)
}
\arguments{
\item{cache_dir}{cache directory}

\item{max_age_days}{maximum age in days before cleaning}

\item{dry_run}{if TRUE, just show what would be deleted}
}
\value{
character vector of files that were (or would be) deleted
}
\description{
Clean old compiled models from cache
}
