% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_cor.R
\name{diagnose_residual_cors}
\alias{diagnose_residual_cors}
\title{Get residual correlation diagnostics}
\usage{
diagnose_residual_cors(resid_cor, threshold = 0.3)
}
\arguments{
\item{resid_cor}{residual_cor object}

\item{threshold}{threshold for flagging high correlations}
}
\value{
list with diagnostic information
}
\description{
Get residual correlation diagnostics
}
