% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{facet_fit}
\alias{facet_fit}
\title{Extract facet fit statistics}
\usage{
facet_fit(object, facet, ...)
}
\arguments{
\item{object}{bmfrm_fit object}

\item{facet}{name of facet}

\item{...}{unused}
}
\description{
Extract facet fit statistics
}
