% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stats.R, R/old_codes.R
\name{facet_infit_outfit}
\alias{facet_infit_outfit}
\title{Compute infit and outfit mean square for any facet}
\usage{
facet_infit_outfit(res_df, facet)

facet_infit_outfit(res_df, facet)
}
\arguments{
\item{res_df}{tibble from extract_residuals()}

\item{facet}{facet column to group by; can be a bare name or a string
e.g. RaterID, "RaterID", ExamineeID, "TaskID", "CriterionID", ...}
}
\value{
tibble with facet_value, model, n_obs, infit_msq, outfit_msq

tibble with facet_value, model, n_obs, infit_msq, outfit_msq
}
\description{
Compute infit and outfit mean square for any facet

Compute infit and outfit mean square for any facet
}
