% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stats.R
\name{facet_infit_outfit_all}
\alias{facet_infit_outfit_all}
\title{Compute infit/outfit for all facets in residual table}
\usage{
facet_infit_outfit_all(res_df)
}
\arguments{
\item{res_df}{tibble from extract_residuals()}
}
\value{
tibble with all facets' infit/outfit statistics
}
\description{
Compute infit/outfit for all facets in residual table
}
