% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{facet_summary}
\alias{facet_summary}
\title{Extract facet-level summaries}
\usage{
facet_summary(object, facet, probs = c(0.025, 0.5, 0.975), ...)
}
\arguments{
\item{object}{bmfrm_fit object}

\item{facet}{name of facet to summarize}

\item{probs}{quantiles for credible intervals}

\item{...}{unused}
}
\description{
Extract facet-level summaries
}
