% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fair_scores}
\alias{fair_scores}
\title{Extract fair scores}
\usage{
fair_scores(object, summary = TRUE, ...)
}
\arguments{
\item{object}{bmfrm_fit object}

\item{summary}{whether to return summary or observation-level scores}

\item{...}{unused}
}
\description{
Extract fair scores
}
