% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_generation.R
\name{generate_bias_param_decl}
\alias{generate_bias_param_decl}
\title{Generate parameter declarations for bias facets}
\usage{
generate_bias_param_decl(spec)
}
\description{
Generate parameter declarations for bias facets
}
\keyword{internal}
