% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_generation.R
\name{generate_bias_transform}
\alias{generate_bias_transform}
\title{Generate transformed parameters for bias facets}
\usage{
generate_bias_transform(spec)
}
\description{
Generate transformed parameters for bias facets
}
\keyword{internal}
