% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_generation.R
\name{generate_eta_sum}
\alias{generate_eta_sum}
\title{Generate eta sum (linear predictor)}
\usage{
generate_eta_sum(spec, fair = FALSE)
}
\description{
Generate eta sum (linear predictor)
}
\keyword{internal}
