% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_generation.R
\name{generate_facet_data_decl}
\alias{generate_facet_data_decl}
\title{Generate data block declarations for facets}
\usage{
generate_facet_data_decl(spec)
}
\description{
Generate data block declarations for facets
}
\keyword{internal}
