% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_generation.R
\name{generate_facet_param_decl}
\alias{generate_facet_param_decl}
\title{Generate parameter declarations for main facets}
\usage{
generate_facet_param_decl(spec)
}
\description{
Generate parameter declarations for main facets
}
\keyword{internal}
