% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_generation.R
\name{generate_facet_transform}
\alias{generate_facet_transform}
\title{Generate transformed parameters for main facets}
\usage{
generate_facet_transform(spec)
}
\description{
Generate transformed parameters for main facets
}
\keyword{internal}
