% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_generation.R
\name{generate_index_extract}
\alias{generate_index_extract}
\title{Generate index extraction code}
\usage{
generate_index_extract(spec)
}
\description{
Generate index extraction code
}
\keyword{internal}
