% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{get_default_priors}
\alias{get_default_priors}
\title{Get default priors for bmfrm models}
\usage{
get_default_priors(
  facets_main = c("person", "item", "rater"),
  facets_bias = NULL
)
}
\arguments{
\item{facets_main}{vector of main facet names}

\item{facets_bias}{vector of bias facet names}
}
\value{
list of bmfrm_prior objects
}
\description{
Get default priors for bmfrm models
}
