% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{get_facet_labels}
\alias{get_facet_labels}
\title{Get facet labels from data_info}
\usage{
get_facet_labels(data_info, facet)
}
\arguments{
\item{data_info}{output from prepare_data_bmfrm()}

\item{facet}{facet name}
}
\value{
character vector of labels
}
\description{
Get facet labels from data_info
}
