% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{get_model_info}
\alias{get_model_info}
\title{Get model compilation info}
\usage{
get_model_info(model_path)
}
\arguments{
\item{model_path}{path to compiled Stan model}
}
\value{
list with compilation details
}
\description{
Get model compilation info
}
