% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{install_cmdstan_if_needed}
\alias{install_cmdstan_if_needed}
\title{Install CmdStan if needed (wrapper around cmdstanr)}
\usage{
install_cmdstan_if_needed(cores = 2, quiet = FALSE)
}
\arguments{
\item{cores}{number of cores to use for compilation}

\item{quiet}{suppress installation messages}
}
\value{
invisible TRUE if successful
}
\description{
Install CmdStan if needed (wrapper around cmdstanr)
}
