% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_cor.R
\name{plot.residual_cor}
\alias{plot.residual_cor}
\title{Plot method for residual correlations}
\usage{
\method{plot}{residual_cor}(x, ...)
}
\arguments{
\item{x}{residual_cor object}

\item{...}{unused}
}
\description{
Plot method for residual correlations
}
