% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pp_check.bmfrm_fit}
\alias{pp_check.bmfrm_fit}
\title{Posterior predictive check method for bmfrm_fit objects}
\usage{
pp_check.bmfrm_fit(
  object,
  type = c("bars", "stat", "all"),
  ndraws = 200,
  seed = 123,
  ...
)
}
\arguments{
\item{object}{bmfrm_fit object}

\item{type}{type of PPC ("bars", "stat", or "all")}

\item{ndraws}{number of posterior draws}

\item{seed}{random seed}

\item{...}{unused}
}
\description{
Posterior predictive check method for bmfrm_fit objects
}
