% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{prepare_data_bmfrm}
\alias{prepare_data_bmfrm}
\title{Prepare data for bmfrm Stan model}
\usage{
prepare_data_bmfrm(spec, data)
}
\arguments{
\item{spec}{bmfrm_spec object from parse_bmfrm_formula()}

\item{data}{data.frame with original data}
}
\value{
list with stan_data, data_clean, and facet_labels
}
\description{
Prepare data for bmfrm Stan model
}
