% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{print.bmfrm_data_check}
\alias{print.bmfrm_data_check}
\title{Print method for data check results}
\usage{
\method{print}{bmfrm_data_check}(x, ...)
}
\arguments{
\item{x}{bmfrm_data_check object}

\item{...}{unused}
}
\description{
Print method for data check results
}
