% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.bmfrm_fit}
\alias{print.bmfrm_fit}
\title{Print method for bmfrm_fit objects}
\usage{
\method{print}{bmfrm_fit}(x, ...)
}
\arguments{
\item{x}{bmfrm_fit object}

\item{...}{unused}
}
\description{
Print method for bmfrm_fit objects
}
