% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{print.bmfrm_prior}
\alias{print.bmfrm_prior}
\title{Print method for bmfrm_prior}
\usage{
\method{print}{bmfrm_prior}(x, ...)
}
\arguments{
\item{x}{bmfrm_prior object}

\item{...}{unused}
}
\description{
Print method for bmfrm_prior
}
