% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_parsing.R
\name{print.bmfrm_spec}
\alias{print.bmfrm_spec}
\title{Print method for bmfrm_spec}
\usage{
\method{print}{bmfrm_spec}(x, ...)
}
\arguments{
\item{x}{bmfrm_spec object}

\item{...}{unused}
}
\description{
Print method for bmfrm_spec
}
