% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_cor.R
\name{print.residual_cor}
\alias{print.residual_cor}
\title{Print method for residual correlation results}
\usage{
\method{print}{residual_cor}(x, ...)
}
\arguments{
\item{x}{residual_cor object}

\item{...}{unused}
}
\description{
Print method for residual correlation results
}
