% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.bmfrm_fit}
\alias{residuals.bmfrm_fit}
\title{Residuals method for bmfrm_fit objects}
\usage{
\method{residuals}{bmfrm_fit}(object, save_draws = FALSE, ...)
}
\arguments{
\item{object}{bmfrm_fit object}

\item{save_draws}{whether to include posterior draws}

\item{...}{unused}
}
\description{
Residuals method for bmfrm_fit objects
}
