% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc.R
\name{run_ppc}
\alias{run_ppc}
\title{Run posterior predictive checks}
\usage{
run_ppc(fit, y_obs, ndraws = 200, seed = 123)
}
\arguments{
\item{fit}{cmdstanr fit object with y_rep in generated quantities}

\item{y_obs}{observed response data}

\item{ndraws}{number of draws to use for PPC}

\item{seed}{random seed}
}
\value{
list with PPC plots and y_rep subset
}
\description{
Run posterior predictive checks
}
