% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{stancode.bmfrm_fit}
\alias{stancode.bmfrm_fit}
\title{Get Stan code from bmfrm_fit object}
\usage{
stancode.bmfrm_fit(object, ...)
}
\arguments{
\item{object}{bmfrm_fit object}

\item{...}{unused}
}
\description{
Get Stan code from bmfrm_fit object
}
