% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{standata.bmfrm_fit}
\alias{standata.bmfrm_fit}
\title{Get Stan data from bmfrm_fit object}
\usage{
standata.bmfrm_fit(object, ...)
}
\arguments{
\item{object}{bmfrm_fit object}

\item{...}{unused}
}
\description{
Get Stan data from bmfrm_fit object
}
