% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summarise_bias}
\alias{summarise_bias}
\title{Summarise bias effects (wrapper)}
\usage{
summarise_bias(object, facet = NULL, prob_thresh = 0.95, ...)
}
\arguments{
\item{object}{bmfrm_fit object}

\item{facet}{bias facet name (if NULL, uses first bias facet)}

\item{prob_thresh}{probability threshold for flagging bias}

\item{...}{unused}
}
\description{
Summarise bias effects (wrapper)
}
