% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.bmfrm_fit}
\alias{summary.bmfrm_fit}
\title{Summary method for bmfrm_fit objects}
\usage{
\method{summary}{bmfrm_fit}(object, facets = NULL, probs = c(0.025, 0.5, 0.975), ...)
}
\arguments{
\item{object}{bmfrm_fit object}

\item{facets}{which facets to summarize (default: all main facets)}

\item{probs}{quantiles for credible intervals}

\item{...}{unused}
}
\description{
Summary method for bmfrm_fit objects
}
