% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{validate_priors}
\alias{validate_priors}
\title{Validate a list of priors}
\usage{
validate_priors(priors, facets_main, facets_bias = NULL)
}
\arguments{
\item{priors}{list of bmfrm_prior objects or NULL}

\item{facets_main}{vector of main facet names}

\item{facets_bias}{vector of bias facet names}
}
\value{
validated list of priors
}
\description{
Validate a list of priors
}
\keyword{internal}
