% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_parsing.R
\name{validate_spec}
\alias{validate_spec}
\title{Validate bmfrm model specification}
\usage{
validate_spec(spec, data)
}
\arguments{
\item{spec}{bmfrm_spec object}

\item{data}{data.frame}
}
\value{
TRUE if valid (invisibly), otherwise stops with error
}
\description{
Validate bmfrm model specification
}
\keyword{internal}
