% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{validate_stan_code}
\alias{validate_stan_code}
\title{Validate Stan code syntax}
\usage{
validate_stan_code(stan_code)
}
\arguments{
\item{stan_code}{character string with Stan code}
}
\value{
list with is_valid (logical) and any error messages
}
\description{
Validate Stan code syntax
}
